Update README.md
I. Motivation

    To give you experience in using arrays, pointers, structs, enums, and different I/O streams and writing program that takes arguments.

    To let you have fun with an application that is extremely captivating.

II. Introduction

The simple world program we will write for this project simulates a number of creatures running around in a simple square world. The world is an m-by-n two-dimensional grid of squares (The number m represents the height of the grid and the number n represents the width of the grid.). Each creature lives in one of the squares, faces in one of the major compass directions (north, east, south, or west) and belongs to a particular species, which determines how that creature behaves.

fl

la

ho

                        ho
                        	

                        fl

Figure 1. A 4-by-4 grid, which contains five creatures. Two creatures belong to the species flytrap (whose short name is “fl”), two belong to the species hop (whose short name is “ho”), and one belongs to the species landmine (whose short name is “la”). The direction of each creature is represented by the direction of the arrow.

represented by the direction of the arrow. For example, the flytrap at the top row is facing east and the flytrap at the bottom row is facing west.